requirements.txt
torchvision
pytorch
matplotlib
scipy
numpy
